[SC-160] Added possibility to create fileTask and graphTask, both need to be validated

+++ api/src/main/java/com/example/api/dto/response/activity/task/SurveyInfoResponse.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class SurveyInfoResponse {
     private String name;
     private String description;
-    private Integer experience;
+    private Double experience;
 }
